feat: 피드 상세보기 뷰 댓글 기능 구현

+++ app/src/main/java/com/teamwss/websoso/data/mapper/FeedMapper.kt
@@ -53,7 +53,7 @@ fun CommentsResponseDto.toData(): CommentsEntity = CommentsEntity(
 
 fun CommentResponseDto.toData(): CommentEntity = CommentEntity(
     user = UserEntity(
-        id = userId.toLong(),
+        id = userId,
         nickname = nickname,
         avatarImage = avatarImage,
     ),
